+++ tools/dev/src/generate-dockerfile.ts
